refactor(settings): separate the list and content into components
